python/abstract_finder/requirements.txt
sqlalchemy
python-dotenv
psycopg2-binary
